Test: productService test code 구현

+++ src/main/java/dto/req/ProductRequestDto.java
@@ -7,11 +7,11 @@ import lombok.Getter;
 public class ProductRequestDto {
 
     private String name;
-    private Float price;
+    private Double price;
     private String description;
 
     @Builder
-    public ProductRequestDto(String name, Float price, String description) {
+    public ProductRequestDto(String name, Double price, String description) {
         this.name = name;
         this.price = price;
         this.description = description;
