+++ src/main/java/entity/Product.java
@@ -22,19 +22,19 @@ public class Product {
 
     private String name;
 
-    private Float price;
+    private Double price;
 
     private String description;
 
     @Builder
-    public Product(Long id, String name, Float price, String description) {
+    public Product(Long id, String name, Double price, String description) {
         this.id = id;
         this.name = name;
         this.price = price;
         this.description = description;
     }
 
-    public void update(String name, Float price, String description) {
+    public void update(String name, Double price, String description) {
         this.name = name;
         this.price = price;
         this.description = description;
